Move environment variable gateway to sub-module
